breakthrough on AI video play
basically we cannot replace a div id with a video, with another one. we have to use separate ids. so with a div id for vid0, vid1, vid2, vid3, we can do the AI update. next problem is that we are continuously triggering mission 2 now, need to check the clearInterval function in updateDB.js